% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_bootstrap.R
\name{AR_innov_sample}
\alias{AR_innov_sample}
\title{Empirical sampling of innovations (errors) provided a fitted stats::ar model. Used in the ARsieve_bootstrap procedure.}
\usage{
AR_innov_sample(ar.fit)
}
\arguments{
\item{ar.fit}{an AR-model fit through one of the stats::ar methods.}
}
\description{
Returns a function from which the innovations are sampled
}
\examples{
series <- arima.sim(model=list(ar=c(0.6, -0.4)), n=100)
ar.fit <- ar.yw(series, max.order=2)
innov.fn <- AR_innov_sample(ar.fit)
innov.fn(150)
}
