% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_bootstrap.R
\name{ARsieve_bootstrap}
\alias{ARsieve_bootstrap}
\title{AR-sieve bootstrap}
\usage{
ARsieve_bootstrap(series, statistic, B, sim.length = NROW(series),
  max.p = 10 * log10(NROW(series)), ar.method = "yw", ...)
}
\arguments{
\item{series}{A univariate linear stationary time series from which the AR-sieve sampling will be done.}

\item{statistic}{Function: the statistic for which a bootstrap-distribution will be created.}

\item{B}{Amount of bootstrap samples.}

\item{sim.length}{Length of the sampled series. Will usually be the same length as the original series,
but can in some cases (such as residual bootstrapping) be useful to be set to another value.}

\item{max.p}{Maximum order the fitted AR model can take.}

\item{ar.method}{Method used to fit the auto-regressive process. ()}

\item{...}{Extra values passed on to the statistic function.}
}
\description{
Bootstrapping method for linear stationary univariate time series. Returns a bootstrap distribution of the chosen statistic.
}
\details{
The implementation follows the AR-sieve bootstrap method as presented by Bühlmann (1997).
Fits an AR process based on AIC minimization and creates samples based on fitted coefficients
and empirical residuals. The statistic is then evaluated on these new samples. Can for example
be used to estimate the variance or confidence intervals of a statistic (e.g. auto-regressive
coefficients of the series).
}
\examples{
series <- arima.sim(model=list(ar=c(0.6, -0.4)), n=100)
ars.res <- ARsieve_bootstrap(series, mean, B=500)
boxplot(ars.res)
}
\references{
Bühlmann, Peter (1997) "Sieve bootstrap for time series".
\emph{Bernoulli}, \bold{3}(2), 123--148.
}
\seealso{
\code{\link{blockwise_bootstrap}}, \code{\link{stats::ar}}
}
\keyword{AR-sieve}
\keyword{ARsieve}
\keyword{bootstrap}
\keyword{sieve}
