% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracdiff.R
\name{backshift}
\alias{backshift}
\title{Backshift a time series}
\usage{
backshift(x, k)
}
\arguments{
\item{x}{Numeric vector. Input time series.}

\item{k}{Non-negative integer. Number of time steps by which to shift.}
}
\value{
Numeric vector of the same length as \code{x}.
}
\description{
Shifts a vector by \code{k} time steps and inserts zeros at the beginning.
}
\keyword{internal}
