% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracdiff.R
\name{binomial_coefficient}
\alias{binomial_coefficient}
\title{Generalized binomial coefficient}
\usage{
binomial_coefficient(r, k)
}
\arguments{
\item{r}{Numeric. Upper argument of the binomial coefficient.}

\item{k}{Non-negative integer. Lower argument of the binomial coefficient.}
}
\value{
Numeric value.
}
\description{
Computes the generalized binomial coefficient used in the expansion of the
fractional differencing operator.
}
\keyword{internal}
