% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_bootstrap.R
\name{block_BS_sample}
\alias{block_BS_sample}
\title{Block-wise bootstrap sample}
\usage{
block_BS_sample(series, l, sim.length = NROW(series))
}
\arguments{
\item{series}{A stationary time series from which the block-wise sampling will be done. Can be univariate or multivariate.}

\item{l}{Length of the blocks}

\item{sim.length}{Length of the new, sampled series}
}
\description{
Gets a single sample according to the block-wise bootstrap scheme. Used in the blockwise_bootstrap function.
}
\examples{
series <- arima.sim(model=list(ar=c(0.6, -0.4)), n=100)
block_BS_sample(series, 5)
}
