% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_bootstrap.R
\name{blockwise_bootstrap}
\alias{blockwise_bootstrap}
\title{Block-wise bootstrap}
\usage{
blockwise_bootstrap(series, statistic, B, l, sim.length = NROW(series),
  ...)
}
\arguments{
\item{series}{A stationary time series from which the block-wise sampling will be done.
Can be univariate or multivariate.}

\item{statistic}{Function: the statistic for which a bootstrap-distribution will be created.}

\item{B}{Amount of bootstrap samples. Choice depends on statistic and purpose}

\item{l}{Length of blocks in the procedure. Good values depend on the series and the statistic,
however, good default values ... n^(1/3) (ref Künsch) or 2n^(1/3) ref (Bühlmann).}

\item{sim.length}{Length of the sampled series. Will usually be the same length as the original series,
but can in some cases (such as residual bootstrapping) be useful to be set to another value.}

\item{...}{Additional values passed on to the statistic function.}
}
\description{
Procedure for evaluating a statistic on a stationary time series.
}
\details{
Returns a bootstrap distribution of the chosen statistic.

The implementation follows the block-wise bootstrap method as presented by Künsch (1989).
Resamples blocks of the time series and uses these to create new time series, which the
statistic is evaluated on. Can for example be used to estimate the variance or confidence
intervals of a statistic (e.g. mean of the series).

Note that the implemented method is not the 'block of blocks'-bootstrap but the more general
'naive' block-wise bootstrap method.
}
\examples{
series <- arima.sim(model=list(ar=c(0.6, -0.4)), n=100)
bw.res <- blockwise_bootstrap(series, mean, B=500, l=5)
boxplot(bw.res)
}
\references{
Kunsch, Hans R. (1989) "The Jackknife and the Bootstrap for General Stationary Observations".
\emph{The Annals of Statistics}, \bold{17}(3), 1217--1241.
}
\seealso{
\code{\link{ARsieve_bootstrap}}
}
\keyword{block}
\keyword{blockwise}
\keyword{bootstrap}
