% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CO2_global}
\alias{CO2_global}
\title{Global monthly mean atmospheric CO2 concentration}
\format{
A data frame with 564 rows and 2 variables:
\describe{
\item{time}{Numeric. Time expressed as fractional year
(for example, 1979.042 for approximately January 1979).}
\item{CO2}{Numeric. Global monthly mean atmospheric CO2 concentration,
in parts per million (ppm).}
}
}
\source{
Obtained from ...
}
\usage{
CO2_global
}
\description{
A monthly time series of global mean atmospheric carbon dioxide (CO2)
concentration values, expressed in parts per million (ppm).
The dataset spans from 1979 to 2025 and is indexed by fractional year.
}
\details{
The dataset contains monthly observations of global atmospheric CO2
concentration.
}
\keyword{datasets}
