% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend.R
\name{detrend}
\alias{detrend}
\title{Linearly detrends a time series}
\usage{
detrend(x)
}
\arguments{
\item{x}{A numerical vector}
}
\value{
Residuals, a vector of the same length as x
}
\description{
Fits a linear regression and returns the residuals
}
