% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hurst.R
\name{.seq_equal_space}
\alias{.seq_equal_space}
\title{Generate equally spaced values}
\usage{
.seq_equal_space(n, min = 0, max = 1)
}
\arguments{
\item{n}{Integer. Number of values.}

\item{min}{Numeric. Lower bound.}

\item{max}{Numeric. Upper bound.}
}
\value{
Numeric vector of equally spaced values.
}
\description{
Internal helper that generates \code{n} equally spaced values between \code{min}
and \code{max}.
}
\keyword{internal}
