% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hurst.R
\name{.seq_log_space}
\alias{.seq_log_space}
\title{Generate logarithmically spaced values}
\usage{
.seq_log_space(n, min = 1, max = exp(1))
}
\arguments{
\item{n}{Integer. Number of values.}

\item{min}{Numeric. Lower bound. Must be positive.}

\item{max}{Numeric. Upper bound. Must be positive.}
}
\value{
Numeric vector of logarithmically spaced values.
}
\description{
Internal helper that generates \code{n} logarithmically spaced values between
\code{min} and \code{max}.
}
\keyword{internal}
