% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hurst.R
\name{estimate_hurst}
\alias{estimate_hurst}
\title{Estimate the Hurst exponent}
\usage{
estimate_hurst(x, method = "rescaled_range", ...)
}
\arguments{
\item{x}{Numeric vector. The time series to analyse.}

\item{method}{Character string. Estimation method. One of
\code{"rescaled_range"} or \code{"aggregated_variance"}.}

\item{...}{Further arguments passed to the selected estimation function.}
}
\value{
Numeric value. Estimated Hurst exponent.
}
\description{
General interface for estimating the Hurst exponent of a univariate time
series. The argument \code{method} selects the estimation method.
}
