% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hurst.R
\name{estimate_hurst_aggvar}
\alias{estimate_hurst_aggvar}
\title{Estimate the Hurst exponent using aggregated variances}
\usage{
estimate_hurst_aggvar(x, sizes = 50, kmin = 10)
}
\arguments{
\item{x}{Numeric vector. The time series to analyse.}

\item{sizes}{Integer. Number of different block sizes.}

\item{kmin}{Integer. Minimum number of blocks required for a block size to
be included.}
}
\value{
Numeric value. Estimated Hurst exponent.
}
\description{
Estimates the Hurst exponent using the aggregated variance method. The slope
of the log-log regression is transformed into a Hurst exponent estimate.
}
