% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hurst.R
\name{estimate_hurst_rs}
\alias{estimate_hurst_rs}
\title{Estimate the Hurst exponent using rescaled range analysis}
\usage{
estimate_hurst_rs(x, windows = 10, sizes = 50, min.size = 10, logkmin = 0.5)
}
\arguments{
\item{x}{Numeric vector. The time series to analyse.}

\item{windows}{Integer. Number of windows sampled for each window size.}

\item{sizes}{Integer. Number of different window sizes.}

\item{min.size}{Integer. Smallest window size considered.}

\item{logkmin}{Numeric. Minimum base-10 log window size included in the
regression.}
}
\value{
Numeric value. Estimated Hurst exponent.
}
\description{
Estimates the Hurst exponent using rescaled range analysis. The Hurst
exponent is estimated as the slope in a log-log regression of the rescaled
range statistic on the window size.
}
