% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fourier_synthetic}
\alias{fourier_synthetic}
\title{Synthetic data for Fourier analysis}
\format{
A data frame with 30 rows and 2 variables:
\describe{
\item{t}{Time index, from 1 to 30.}
\item{x}{Observed synthetic time series value.}
}
}
\source{
Simulated data for teaching purposes.
}
\usage{
fourier_synthetic
}
\description{
A synthetic univariate time series used to illustrate discrete Fourier
decomposition and reconstruction.
}
\details{
The series contains a small number of periodic components and is intended for
teaching the interpretation of Fourier coefficients, dominant frequencies,
and partial reconstruction from selected Fourier terms.
}
\examples{
data(fourier_synthetic)
plot(fourier_synthetic, type = "b", main = "Synthetic Data")

}
\keyword{datasets}
