% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hurst.R
\name{hurst.aggregated.variance}
\alias{hurst.aggregated.variance}
\title{#' Hurst aggregated variances helper}
\usage{
hurst.aggregated.variance(x, sizes = 50, kmin = 10)
}
\description{
#' Hurst aggregated variances helper
}
