% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hurst.R
\name{hurst_aggregated_variance}
\alias{hurst_aggregated_variance}
\title{#' Hurst aggregated variances helper}
\usage{
hurst_aggregated_variance(x, sizes = 50, kmin = 10)
}
\description{
#' Hurst aggregated variances helper
}
