% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_origin.R
\name{MAE}
\alias{MAE}
\title{Mean average error}
\usage{
MAE(object, ...)
}
\arguments{
\item{object}{object of class with beloning MAE function}
}
\description{
Computes the mean average error (MAE) the specified class
}
