% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_origin.R
\name{MAE.ROE_object}
\alias{MAE.ROE_object}
\title{Mean average error for an ROE_object}
\usage{
MAE.ROE_object(object)
}
\arguments{
\item{object}{A rolling origin evaluation object}
}
\description{
Computes the mean average error (MAE) for each forecast horizon
}
