% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_origin.R
\name{MSE}
\alias{MSE}
\title{Mean squared error}
\usage{
MSE(object, ...)
}
\arguments{
\item{object}{object of class with beloning MSE function}
}
\description{
Computes the mean squared error (MSE) the specified class
}
