% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_origin.R
\name{MSE.ROE_object}
\alias{MSE.ROE_object}
\title{Mean squared error for an ROE_object}
\usage{
MSE.ROE_object(object)
}
\arguments{
\item{object}{A rolling origin evaluation object}
}
\description{
Computes the mean-squared-error (RMSE) for each forecast horizon
}
