% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_bootstrap.R
\name{optimal_bootstrap}
\alias{optimal_bootstrap}
\title{A bootstrap procedure where, not suitable for a ...}
\usage{
optimal_bootstrap(sim.fn, statistic, B, sim.length, ...)
}
\arguments{
\item{sim.fn}{Simulation function (distribution) from which the time series is drawn from.}

\item{statistic}{Function: the statistic for which a bootstrap-distribution will be created.}

\item{B}{Amount of bootstrap samples.}

\item{sim.length}{Length of the sampled series. Will usually be the same length as the original series,
but can in some cases (such as residual bootstrapping) be useful to be set to another value.}

\item{...}{Extra values passed on to the statistic function.}
}
\description{
A bootstrap procedure where, not suitable for a ...
}
\examples{
series <- arima.sim(model=list(ar=c(0.6, -0.4)), n=100)
block_BS_sample(series, arma_coef_stat, 5, )
}
\keyword{block}
\keyword{blockwise}
\keyword{bootstrap}
\keyword{series}
\keyword{time}
