% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hurst.R
\name{plot.hurst.aggregated.variance}
\alias{plot.hurst.aggregated.variance}
\title{Plot Hurst aggregated variances}
\usage{
\method{plot}{hurst.aggregated.variance}(x)
}
\description{
Plot Hurst aggregated variances
}
