% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fourier.R
\name{plot_fourier_approximation}
\alias{plot_fourier_approximation}
\title{Plot a Fourier approximation}
\usage{
plot_fourier_approximation(
  x,
  g1 = zero_function,
  g2 = zero_function,
  g3 = zero_function,
  g4 = zero_function,
  g5 = zero_function,
  ...
)
}
\arguments{
\item{x}{Numeric vector or time series. Original series.}

\item{g1}{Function. First Fourier component function.}

\item{g2}{Function. Second Fourier component function.}

\item{g3}{Function. Third Fourier component function.}

\item{g4}{Function. Fourth Fourier component function.}

\item{g5}{Function. Fifth Fourier component function.}

\item{...}{Further arguments passed to \code{plot()}.}
}
\value{
No return value. Called for its side effect of producing a plot.
}
\description{
Plots a time series and overlays an approximation constructed from selected
Fourier component functions.
}
