% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fourier.R
\name{plot_fourier_orthogonality}
\alias{plot_fourier_orthogonality}
\title{Plot Fourier orthogonality}
\usage{
plot_fourier_orthogonality(f, g, kmax = 3)
}
\arguments{
\item{f}{Function. First trigonometric function, usually \code{sin} or \code{cos}.}

\item{g}{Function. Second trigonometric function, usually \code{sin} or \code{cos}.}

\item{kmax}{Integer. Maximum frequency shown.}
}
\value{
No return value. Called for its side effect of producing a plot.
}
\description{
Visualizes the orthogonality of sine and cosine functions for different
frequencies.
}
