% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracdiff.R
\name{plot_parameter_grid}
\alias{plot_parameter_grid}
\title{Plot a function for nine parameter values}
\usage{
plot_parameter_grid(f, min = 0, max = 1)
}
\arguments{
\item{f}{Function. A plotting function that takes one numeric argument.}

\item{min}{Numeric. Lower bound of the parameter range.}

\item{max}{Numeric. Upper bound of the parameter range.}
}
\value{
No return value. Called for its side effect of producing a 3 by 3
panel of plots.
}
\description{
Evaluates a plotting function for nine equally spaced parameter values.
}
\details{
This function is useful for comparing how a plot changes as a parameter
varies. In this exercise, it is used to compare different values of the
fractional differencing parameter \code{d}.
}
\examples{
plot_parameter_grid(plot_fractional_difference_coefficients, min = 0, max = 1)

}
