% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend.R
\name{plot_trend}
\alias{plot_trend}
\title{Plot a time series and its linear trend}
\usage{
plot_trend(t, x)
}
\arguments{
\item{t}{A vector of time steps plotted on x axis}

\item{x}{A numerical vector plotted on y axis}
}
\value{
No return value, called for side effects (a plot).
}
\description{
Fits a linear regression and visualizes the time series and the linear trend
}
