% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{plotcwt}
\alias{plotcwt}
\title{Plot cwt}
\usage{
plotcwt(
  cw,
  nlevels = 20,
  contour = T,
  shortage = max(1, ceiling(nrow(cw)/1000)),
  shortop = mean,
  main = "Wavelet Powerspectrum",
  ar = NULL,
  palette = pal,
  tunit = "years",
  rescale.method = "quantile",
  ...
)
}
\description{
Plot cwt
}
