% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_origin.R
\name{RMSE}
\alias{RMSE}
\title{Root mean squared error}
\usage{
RMSE(object, ...)
}
\arguments{
\item{object}{object of class with beloning RMSE function}
}
\description{
Computes the root-mean-squared-error (RMSE) the specified class
}
