% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_origin.R
\name{RMSE.ROE_object}
\alias{RMSE.ROE_object}
\title{Root mean squared error for an ROE_object}
\usage{
RMSE.ROE_object(object)
}
\arguments{
\item{object}{A rolling origin evaluation object}
}
\description{
Computes the root-mean-squared-error (RMSE) for each forecast horizon
}
