% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_origin.R
\name{rolling_origin_eval}
\alias{rolling_origin_eval}
\title{Rolling origin forecast evaluation}
\usage{
rolling_origin_eval(series, start, forecast_fn, h = 1, ...)
}
\arguments{
\item{series}{Univariate time series used for fitting and computing forecast errors.}

\item{start}{Time point used as the starting point for the rolling origin forecast.}

\item{forecast_fn}{Function which returns forecasts. Takes a time series as its first argument and has an argument
h representing the forecast horizon. Returns a vector of length equal to the given forecast horizon h.}

\item{h}{Forecast horizon. This function will evaluate all horizons 1 to h.}

\item{...}{Additional arguments passed to forecast_fn}
}
\description{
Rolling origin forecast evaluation, a.k.a. time-series cross validation, of a model or method.
Computes errors and prediction of a forecast function applied to a time series according to the rolling origin scheme.
}
\details{
This method implements the rolling origin forecast evaluation (see e.g. Hyndman and Athanasopoulos, 2018).
Returns a list of two matrices (multivariate time series), one containing the errors and one the predictions,
where columns represent forecast horizon and rows represent time points.

The method starts with a subset X[1:t] and forecasts X[(t+1):(t+h)] based on this subset.
Then forecasts X[(t+2):(t+h+1)] based on X[1:(t+1)] and so on.
}
\examples{
test(1, 2)
}
\references{
Hyndman, R.J., & Athanasopoulos, G. (2018)
\emph{Forecasting: principles and practice, 2nd edition}, OTexts: Melbourne, Australia.
OTexts.com/fpp2. Accessed on 2019-10-18
}
\keyword{evaluation}
\keyword{origin}
\keyword{rolling}
