% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_origin.R
\name{rolling_origin_forecast}
\alias{rolling_origin_forecast}
\title{Rolling origin forecast evaluation, a.k.a. time-series cross validation.}
\usage{
rolling_origin_forecast(series, start, fn, h = 1, ...)
}
\arguments{
\item{series}{series used for ...}

\item{start}{starting origin for the rolling origin forecast}

\item{fn}{forecasting function of the series}

\item{h}{forecasting horizon. Will evaluate all ... from 1 to h}
}
\description{
This function does something
}
\examples{
test(1, 2)
}
\keyword{evaluation}
\keyword{origin}
\keyword{rolling}
