% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_origin.R
\name{test}
\alias{test}
\title{A test function}
\usage{
test(a, b, ...)
}
\arguments{
\item{a}{integer wallah}

\item{b}{integer brah}
}
\description{
This function allows you to express your love of cats.
}
\examples{
test(1, 2)
}
\keyword{add}
