% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{waveheight_Draupner}
\alias{waveheight_Draupner}
\title{Draupner wave height time series}
\format{
A time series object of class \code{\link{ts}}.
\describe{
\item{Time series start}{1}
\item{Time series end}{3.34375}
\item{Frequency}{2.133333}
\item{Values}{Wave-height measurements, in metres if stored in the original Draupner units.}
}
}
\source{
Draupner platform wave-height record.
}
\usage{
waveheight_Draupner
}
\description{
A univariate time series of wave-height measurements from the Draupner
platform in the North Sea.
}
\details{
The object is stored as a \code{\link{ts}} time-series object.
}
\examples{
data(waveheight_Draupner)
plot(waveheight_Draupner)

}
\keyword{datasets}
