% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weather_Hurdal}
\alias{weather_Hurdal}
\title{Weather data from Hurdal, Norway}
\format{
A data frame with 17520 rows and 4 variables:
\describe{
\item{weather_Hurdal}{date, POSIXct}
\item{Precipitation (mm)}{Numeric observed value}
\item{Tair@42m (deg C)}{Numeric observed value}
\item{Tair@2m (deg C)}{Numeric observed value}
}
}
\source{
Holger Lange
}
\usage{
weather_Hurdal
}
\description{
A dataset with weather data recorded in Hurdal, Norway.
}
\keyword{datasets}
