% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fourier.R
\name{zero_function}
\alias{zero_function}
\title{Return a zero-valued function}
\usage{
zero_function(t)
}
\arguments{
\item{t}{Numeric input.}
}
\value{
Numeric value 0.
}
\description{
Internal helper that returns zero for any input.
}
\keyword{internal}
